src/agent/gym_node/src/planning/requirement.txt
cpprb==10.1.1
easydict==1.9
matplotlib
gym
joblib
numpy
pandas
pillow
pynput==1.7.3
pyyaml==5.4.1
scikit-image
scipy
seaborn==0.11.2
tabulate==0.8.9
tensorboard>=2.7.0
tensorboard-data-server>=0.6.1
tensorboard-plugin-wit>=1.8.0
tensorboardx>=2.4
ray
ray[tune]
tqdm~=4.62.3
DI-engine
lmdb